add Qcov7 into common pool as a likert scale question
