+++ src/editor-example-generators/italy/inf-weekly.ts
@@ -1,10 +1,10 @@
 import { Survey, SurveyGroupItem, SurveyItem } from "survey-engine/lib/data_types";
 import { ItemEditor } from "../../editor-engine/survey-editor/item-editor";
 import { SurveyEditor } from "../../editor-engine/survey-editor/survey-editor";
-import { initDropdownGroup, initLikertScaleItem, initMultipleChoiceGroup, initSingleChoiceGroup } from "../../editor-engine/utils/question-type-generator";
+import { initDropdownGroup, initSingleChoiceGroup } from "../../editor-engine/utils/question-type-generator";
 import { expWithArgs, generateHelpGroupComponent, generateLocStrings, generateTitleComponent } from "../../editor-engine/utils/simple-generators";
 import { WeeklyQuestions as CommonPoolWeekly } from "../common_question_pool/influenzanet-weekly";
-import { likertScaleKey, multipleChoiceKey, responseGroupKey, singleChoiceKey } from "../common_question_pool/key-definitions";
+import { multipleChoiceKey, responseGroupKey, singleChoiceKey } from "../common_question_pool/key-definitions";
 
 const weekly = (): Survey | undefined => {
     const surveyKey = 'weekly';
@@ -192,6 +192,10 @@ const weekly = (): Survey | undefined => {
     const Q_dailyRoutineDaysMissed = CommonPoolWeekly.dailyRoutineDaysMissed(hasSymptomGroupKey, Q_dailyRoutine.key, true);
     survey.addExistingSurveyItem(Q_dailyRoutineDaysMissed, hasSymptomGroupKey);
 
+    // // Qcov7 Covid 19 habits change question ------------------------------------------------------
+    const Q_covidHabits = CommonPoolWeekly.covidHabitsChange(hasSymptomGroupKey, false);
+    survey.addExistingSurveyItem(Q_covidHabits, hasSymptomGroupKey);
+
     // // Q_BE_7x duration hospitalisation----------------------------------------------
     const Q_durHosp = durHosp(hasSymptomGroupKey, Q_visitedMedicalService.key, true, "Q_BE_7x");
     survey.addExistingSurveyItem(Q_durHosp, hasSymptomGroupKey);
@@ -207,10 +211,6 @@ const weekly = (): Survey | undefined => {
     const Q_causeOfSymptoms = CommonPoolWeekly.causeOfSymptoms(hasSymptomGroupKey, true);
     survey.addExistingSurveyItem(Q_causeOfSymptoms, hasSymptomGroupKey);
 
-    // // Qcov_BE_7 Covid 19 habits change question ------------------------------------------------------
-    const Q_covidHabits = covidHabitsChange(hasSymptomGroupKey, false, "Qcov_BE_7");
-    survey.addExistingSurveyItem(Q_covidHabits, hasSymptomGroupKey);
-
     const surveyEndText = surveyEnd(rootKey);
     survey.addExistingSurveyItem(surveyEndText, rootKey);
 
@@ -218,327 +218,6 @@ const weekly = (): Survey | undefined => {
 }
 
 export default weekly;
-
-/**
- * TEST COVID-19: single choice question about COVID-19  test
- *
- * @param parentKey full key path of the parent item, required to genrate this item's unique key (e.g. `<surveyKey>.<groupKey>`).
- * @param isRequired if true adds a default "hard" validation to the question to check if it has a response.
- * @param keyOverride use this to override the default key for this item (only last part of the key, parent's key is not influenced).
- */
-const covidTest = (parentKey: string, isRequired?: boolean, keyOverride?: string): SurveyItem => {
-    const defaultKey = 'Q_BE_1_3'
-    const itemKey = [parentKey, keyOverride ? keyOverride : defaultKey].join('.');
-    const editor = new ItemEditor(undefined, { itemKey: itemKey, isGroup: false });
-    editor.setVersion(1);
-
-    // QUESTION TEXT
-    editor.setTitleComponent(
-        generateTitleComponent(new Map([
-            ["nl-be", "Heeft u sinds de vorige vragenlijst een COVID-19 test laten uitvoeren (of in de afgelopen 7 dagen indien dit uw eerste bezoek is)?"],
-            ["fr-be", "Avez-vous passé un test de dépistage du coronavirus depuis le dernier questionnaire (ou au cours des 7 derniers jours s'il s'agit de votre première visite) ?"],
-            ["de-be", "Haben Sie seit der letzten Fragen einen COVID-19-Test durchführen lassen (oder in den vergangenen 7 Tagen, wenn dieser Ihr erster Besuch ist)?"],
-            ["en", "Have you been tested for coronavirus since the last questionnaire (or within the past 7 days if this is your first visit)?"],
-        ]))
-    );
-
-    // CONDITION
-    // none
-
-    // INFO POPUP
-    editor.setHelpGroupComponent(
-        generateHelpGroupComponent([
-            {
-                content: new Map([
-                    ["nl-be", "Waarom vragen we dit?"],
-                    ["fr-be", "Pourquoi posons-nous cette question ?"],
-                    ["de-be", "Warum fragen wir das?"],
-                    ["en", "Why are we asking this question?"],
-                ]),
-                style: [{ key: 'variant', value: 'h5' }],
-            },
-            {
-                content: new Map([
-                    ["nl-be", "We willen weten hoe COVID-19 zich verspreidt in de bevolking."],
-                    ["fr-be", "Nous voulons savoir comment le coronavirus se propage au sein de la population."],
-                    ["de-be", "Wir möchten wissen, wie Covid-19 sich in der Bevölkerung verbreitet."],
-                    ["en", "We want to understand how the coronavirus is spreading within the population."],
-                ]),
-                //style: [{ key: 'variant', value: 'p' }],
-            },
-        ])
-    );
-
-    // RESPONSE PART
-    const rg = editor.addNewResponseComponent({ role: 'responseGroup' });
-    const rg_inner = initSingleChoiceGroup(singleChoiceKey, [
-        {
-            key: '0', role: 'option', content: new Map([
-                ["nl-be", "Ja"],
-                ["fr-be", "Oui"],
-                ["de-be", "Ja"],
-                ["en", "Yes"],
-
-            ])
-        },
-        {
-            key: '1', role: 'option',
-            content: new Map([
-                ["nl-be", "Nee"],
-                ["fr-be", "Non"],
-                ["de-be", "Nein"],
-                ["en", "No"],
-            ])
-        },
-        {
-            key: '2', role: 'option',
-            content: new Map([
-                ["nl-be", "Nee, ik weet het niet (meer)"],
-                ["fr-be", "Je ne sais pas (plus)"],
-                ["de-be", "Ich weiß es nicht (mehr)"],
-                ["en", "I don’t know/can’t remember"],
-            ])
-        },
-    ]);
-    editor.addExistingResponseComponent(rg_inner, rg?.key);
-
-    // VALIDATIONs
-    if (isRequired) {
-        editor.addValidation({
-            key: 'r1',
-            type: 'hard',
-            rule: expWithArgs('hasResponse', itemKey, responseGroupKey)
-        });
-    }
-
-    return editor.getItem();
-}
-
-/**
- * REASON COVID-19 test: single choice question about COVID-19  test
- *
- * @param parentKey full key path of the parent item, required to genrate this item's unique key (e.g. `<surveyKey>.<groupKey>`).
- * @param isRequired if true adds a default "hard" validation to the question to check if it has a response.
- * @param keyOverride use this to override the default key for this item (only last part of the key, parent's key is not influenced).
- */
-const reasonTest = (parentKey: string, keycovidTest?: string, isRequired?: boolean, keyOverride?: string): SurveyItem => {
-    const defaultKey = 'Q_BE_1_4'
-    const itemKey = [parentKey, keyOverride ? keyOverride : defaultKey].join('.');
-    const editor = new ItemEditor(undefined, { itemKey: itemKey, isGroup: false });
-    editor.setVersion(1);
-
-    // QUESTION TEXT
-    editor.setTitleComponent(
-        generateTitleComponent(new Map([
-            ["nl-be", "Wat was de reden dat u een COVID-19 test liet uitvoeren?"],
-            ["fr-be", "Pour quelle raison avez-vous effectué un test de dépistage du coronavirus ?"],
-            ["de-be", "Was war der Grund dafür, dass Sie einen COVID-19-Test durchführen ließen?"],
-            ["en", "Why are you being tested for coronavirus?"],
-        ]))
-    );
-
-    // CONDITION
-    if (keycovidTest) {
-        editor.setCondition(
-            expWithArgs('responseHasKeysAny', keycovidTest, [responseGroupKey, singleChoiceKey].join('.'), '0')
-        );
-    }
-
-    // INFO POPUP
-    editor.setHelpGroupComponent(
-        generateHelpGroupComponent([
-            {
-                content: new Map([
-                    ["nl-be", "Hoe moet ik deze vraag beantwoorden?"],
-                    ["fr-be", "Comment dois-je répondre à cette question ?"],
-                    ["de-be", "Wie soll ich diese Frage beantworten?"],
-                    ["en", "How should I answer this question?"],
-                ]),
-                style: [{ key: 'variant', value: 'h5' }],
-            },
-            {
-                content: new Map([
-                    ["nl-be", "Antwoord zo precies mogelijk."],
-                    ["fr-be", "Répondez de la manière la plus précise possible."],
-                    ["de-be", "Bitte antworten Sie so genau wie möglich."],
-                    ["en", "Answer as precisely as possible."],
-                ]),
-                //style: [{ key: 'variant', value: 'p' }],
-            },
-        ])
-    );
-
-    // RESPONSE PART
-    const rg = editor.addNewResponseComponent({ role: 'responseGroup' });
-    editor.addExistingResponseComponent({
-        role: 'text',
-        style: [{ key: 'className', value: 'mb-2' }],
-        content: generateLocStrings(
-            new Map([
-                ['nl-be', 'Meerdere antwoorden mogelijk'],
-                ["fr-be", "Plusieurs réponses sont possibles"],
-                ["de-be", "Mehrere Antworten möglich"],
-                ["en", "Multiple answers possible"],
-            ])),
-    }, rg?.key);
-    const rg_inner = initMultipleChoiceGroup(multipleChoiceKey, [
-        {
-            key: '0', role: 'option', content: new Map([
-                ["nl-be", "Vanwege mijn symptomen"],
-                ["fr-be", "En raison de mes symptômes"],
-                ["de-be", "Aufgrund meiner Symptome"],
-                ["en", "Because of my symptoms"],
-            ])
-        },
-        {
-            key: '1', role: 'option',
-            content: new Map([
-                ["nl-be", "Vanwege symptomen of een positieve test bij één van mijn huisgenoten"],
-                ["fr-be", "En raison de symptômes ou d'un test positif auprès de l'une des personnes vivant avec moi"],
-                ["de-be", "Aufgrund von Symptomen oder einem positiven Test bei einem meiner Mitbewohner (Hausgenossen)"],
-                ["en", "Because a person who lives with me experienced symptoms or had a positive coronavirus test"],
-            ])
-        },
-        {
-            key: '2', role: 'option',
-            content: new Map([
-                ["nl-be", "Vanwege symptomen of een positieve test bij één van mijn contacten buitenshuis"],
-                ["fr-be", "En raison de symptômes ou d'un test positif auprès de l'un de mes contacts extérieurs"],
-                ["de-be", "Aufgrund von Symptomen oder einem positiven Test bei einem meiner Kontakte außerhalb des Hauses"],
-                ["en", "Because one of my external contacts experienced symptoms or had a positive coronavirus test"],
-            ])
-        },
-        {
-            key: '3', role: 'option',
-            content: new Map([
-                ["nl-be", "Vanwege mijn terugkeer uit een risicogebied"],
-                ["fr-be", "En raison de mon retour d'une zone à risque"],
-                ["de-be", "Aufgrund meiner Rückkehr aus einem Risikogebiet"],
-                ["en", "Because I returned from a risk area"],
-            ])
-        },
-        {
-            key: '4', role: 'option',
-            content: new Map([
-                ["nl-be", "Vanwege mijn deelname aan een risico-activiteit (vliegreis, evenement, ...)"],
-                ["fr-be", "En raison de ma participation à une activité à risque (un voyage en avion, un événement, etc.)"],
-                ["de-be", "Aufgrund meiner Teilnahme an einer Risiko-Aktivität (Flugreise, Veranstaltung, Event usw.)"],
-                ["en", "Because I took part in a high-risk activity (a plane trip, an event, etc.)"],
-            ])
-        },
-        {
-            key: '5', role: 'option',
-            content: new Map([
-                ["nl-be", "Vanwege mijn beroep"],
-                ["fr-be", "En raison de ma profession"],
-                ["de-be", "Aufgrund meines Berufs"],
-                ["en", "Because of my profession"],
-            ])
-        },
-        {
-            key: '6', role: 'input',
-            style: [{ key: 'className', value: 'w-100' }],
-            content: new Map([
-                ["nl-be", "Andere"],
-                ["fr-be", "Autre"],
-                ["de-be", "Andere"],
-                ["en", "Other"],
-            ]),
-            description: new Map([
-                ["nl-be", "Beschrijf hier (optioneel in te vullen)"],
-                ["fr-be", "Veuillez fournir une description ici (facultatif)"],
-                ["de-be", "Beschreiben Sie es hier (optional einzutragen)"],
-                ["en", "Describe here (optional)"],
-            ])
-        },
-    ]);
-    editor.addExistingResponseComponent(rg_inner, rg?.key);
-
-    // VALIDATIONs
-    if (isRequired) {
-        editor.addValidation({
-            key: 'r1',
-            type: 'hard',
-            rule: expWithArgs('hasResponse', itemKey, responseGroupKey)
-        });
-    }
-
-    return editor.getItem();
-}
-
-/**
- * DATE COVID-19 test: date COVID-19 test
- *
- * @param parentKey full key path of the parent item, required to genrate this item's unique key (e.g. `<surveyKey>.<groupKey>`).
- * @param isRequired if true adds a default "hard" validation to the question to check if it has a response.
- * @param keyOverride use this to override the default key for this item (only last part of the key, parent's key is not influenced).
- */
-const dateTest = (parentKey: string, keycovidTest?: string, isRequired?: boolean, keyOverride?: string): SurveyItem => {
-    const defaultKey = 'Q_BE_1_5'
-    const itemKey = [parentKey, keyOverride ? keyOverride : defaultKey].join('.');
-    const editor = new ItemEditor(undefined, { itemKey: itemKey, isGroup: false });
-    editor.setVersion(1);
-
-    // QUESTION TEXT
-    editor.setTitleComponent(
-        generateTitleComponent(new Map([
-            ["nl-be", "Wanneer heeft u een test laten uitvoeren voor COVID-19? Als u de datum niet meer precies weet, mag u deze schatten."],
-            ["fr-be", "Quand avez-vous fait réaliser un test de dépistage du coronavirus ? Si vous ne vous souvenez pas (plus) de la date exacte, vous pouvez l'estimer."],
-            ["de-be", "Wann haben Sie einen Test auf COVID-19 durchführen lassen? Wenn Sie das Datum nicht mehr genau wissen, dürfen Sie es schätzen."],
-            ["en", "When did you get tested for the coronavirus? If you don't remember the exact date, please give an estimate."],
-        ]))
-    );
-
-    // CONDITION
-    if (keycovidTest) {
-        editor.setCondition(
-            expWithArgs('responseHasKeysAny', keycovidTest, [responseGroupKey, singleChoiceKey].join('.'), '0')
-        );
-    }
-
-    // INFO POPUP
-    //
-
-    // RESPONSE PART
-    const rg = editor.addNewResponseComponent({ role: 'responseGroup' });
-    const rg_inner = initSingleChoiceGroup(singleChoiceKey, [
-        {
-            key: '0', role: 'dateInput',
-            optionProps: {
-                min: { dtype: 'exp', exp: expWithArgs('timestampWithOffset', -5184000) },
-                max: { dtype: 'exp', exp: expWithArgs('timestampWithOffset', 0) }
-            },
-            description: new Map([
-                ["nl-be", "Kies een datum"],
-                ["fr-be", "Choisissez la date"],
-                ["de-be", "Datum auswählen"],
-                ["en", "Choose date"],
-            ]),
-        },
-        {
-            key: '1', role: 'option',
-            content: new Map([
-                ["nl-be", "Ik weet het niet (meer)"],
-                ["fr-be", "Je ne sais pas (plus)"],
-                ["de-be", "Ich weiß es nicht (mehr)"],
-                ["en", "I don’t know/can’t remember"],
-
-            ])
-        },
-    ]);
-    editor.addExistingResponseComponent(rg_inner, rg?.key);
-
-    // VALIDATIONs
-    if (isRequired) {
-        editor.addValidation({
-            key: 'r1',
-            type: 'hard',
-            rule: expWithArgs('hasResponse', itemKey, responseGroupKey)
-        });
-    }
-
-    return editor.getItem();
-}
 
 /**
  * DURATION HOSPITAL
@@ -1034,334 +713,6 @@ const inComa = (parentKey: string, keyVisitedMedicalServ: string, isRequired?: b
             rule: expWithArgs('hasResponse', itemKey, responseGroupKey)
         });
     }
-
-    return editor.getItem();
-}
-
-
-/**
- * COVID 19 Personal Habits Changes: likert scale question about changes in personal habits after experiencing covid symptoms
- *
- * @param parentKey full key path of the parent item, required to genrate this item's unique key (e.g. `<surveyKey>.<groupKey>`).
- * @param keySymptomsQuestion reference to the symptom survey
- * @param isRequired if true adds a default "hard" validation to the question to check if it has a response.
- * @param keyOverride use this to override the default key for this item (only last part of the key, parent's key is not influenced).
- */
-const covidHabitsChange = (parentKey: string, isRequired?: boolean, keyOverride?: string): SurveyItem => {
-    const defaultKey = 'Qcov_BE_7'
-    const itemKey = [parentKey, keyOverride ? keyOverride : defaultKey].join('.');
-    const editor = new ItemEditor(undefined, { itemKey: itemKey, isGroup: false });
-    editor.setVersion(1);
-
-    // QUESTION TEXT
-    editor.setTitleComponent(
-        generateTitleComponent(new Map([
-            ["nl-be", "Heeft u vanwege uw symtomen deze maatregelen genomen of verstrengd (in vergelijking met de periode voor uw symptomen)?"],
-            ["fr-be", "En raison de vos symptômes, avez-vous pris ou intensifié les mesures reprises ci-dessous (par rapport à la période précédant vos symptômes) ?"],
-            ["de-be", "Haben Sie aufgrund Ihrer Symtome diese Maßnahmen ergriffen oder verstärkt (im Vergleich mit dem Zeitraum vor Ihren Symptomen)?"],
-            ["en", "Did you begin to follow or increase any of the measures below, due to your symptoms (compared to the period before your symptoms began)?"],
-        ]))
-    );
-
-    // CONDITION
-    // none
-
-    // INFO POPUP
-    editor.setHelpGroupComponent(
-        generateHelpGroupComponent([
-            {
-                content: new Map([
-                    ["nl-be", "Waarom vragen we dit?"],
-                    ["fr-be", "Pourquoi posons-nous cette question ?"],
-                    ["de-be", "Warum fragen wir das?"],
-                    ["en", "Why are we asking this question?"],
-                ]),
-                style: [{ key: 'variant', value: 'h5' }],
-            },
-            {
-                content: new Map([
-                    ["nl-be", "Om te onderzoeken hoe verschillende maatregelen worden opgevolgd."],
-                    ["fr-be", "Afin d'examiner la manière dont les différentes mesures sont suivies."],
-                    ["de-be", "Um zu untersuchen, wie verschiedene Maßnahmen befolgt werden."],
-                    ["en", "To examine how different measures are being followed."],
-                ]),
-                style: [{ key: 'variant', value: 'p' }],
-            },
-            {
-                content: new Map([
-                    ["nl-be", "Hoe moet ik deze vraag beantwoorden?"],
-                    ["fr-be", "Comment dois-je répondre à cette question ?"],
-                    ["de-be", "Wie soll ich diese Frage beantworten?"],
-                    ["en", "How should I answer this question?"],
-                ]),
-                style: [{ key: 'variant', value: 'h5' }],
-            },
-            {
-                content: new Map([
-                    ["nl-be", "Bijvoorbeeld 'Handen schudden vermijden': Kies 'Ja', omwille van mijn symptomen heb ik geen of nog minder handen geschud; Kies 'Nee, ik pas deze maatregel niet toe', ik geef nog steeds mensen de hand ook al heb ik symptomen; Kies 'Nee, ik hield me al aan deze maatregel', want ook voor mijn symptomen gaf ik geen handen en heb dit niet aangepast."],
-                    ["fr-be", "Par exemple, « Évitez de serrer les mains » : Choisissez « Oui » si vous avez serré moins de mains ou si vous n'avez plus serré de mains en raison de vos symptômes ; Choisissez « Non, je n'applique pas cette mesure », si vous continuez à serrer la main des gens malgré vos symptômes ; Choisissez « Non, J'appliquais déjà cette mesure » si vous ne serriez déjà plus la main des gens avant l'apparition de vos symptômes, et que vous n'avez pas modifié ce comportement."],
-                    ["de-be", "Zum Beispiel so, 'Hände schütteln vermeiden': Wählen Sie 'Ja', aufgrund meiner Symptome habe ich keine oder noch weniger Hände geschüttelt; Wählen Sie 'Nein, ich wende diese Maßnahme nicht an', ich gebe noch immer Menschen die Hand, obwohl ich schon Symptome habe; Wählen Sie 'Nein, ich hielt mich schon an diese Maßnahme', denn auch aufgrund meiner Symptome gab ich niemandem die Hand und habe das nicht gemacht."],
-                    ["en", "For example, 'Avoid shaking hands': Answer 'yes' if you shake hands less or not at all due to your symptoms; Answer 'No, I am not following this measure' if you continue to shake hands despite your symptoms; Answer 'No, I was already following this measure' if you had already stopped shaking hands before the onset of your symptoms and you did not change this behaviour."],
-                ]),
-                // style: [{ key: 'variant', value: 'p' }],
-            },
-        ])
-    );
-
-
-    // RESPONSE PART
-    const likertOptions = [
-        {
-            key: "1", content: new Map([
-                ["nl-be", "Ja, ik pas deze maatregel nu (voor het eerst) of meer (verstrengd) toe"],
-                ["fr-be", "Oui, j'applique cette mesure maintenant (pour la première fois) ou de manière plus stricte"],
-                ["de-be", "Ja, ich wende diese Maßnahme nun (zum ersten Mal) oder mehr (verstärkt) an"],
-                ["en", " Yes, I am following this measure now for the first time, or in a stricter way"],
-            ])
-        },
-        {
-            key: "3", content: new Map([
-                ["nl-be", "Nee, ik hield me al aan deze maatregel"],
-                ["fr-be", "Non, j'appliquais déjà cette mesure"],
-                ["de-be", "Nein, ich hielt mich schon an diese Maßnahme"],
-                ["en", "I was already following this measure"],
-            ])
-        },
-        {
-            key: "2", content: new Map([
-                ["nl-be", "Nee, ik pas deze maatregel niet toe"],
-                ["fr-be", "Non, je n'applique pas cette mesure"],
-                ["de-be", "Nein, ich wende diese Maßnahme nicht an"],
-                ["en", "No, I am not following this measure"],
-            ])
-        }
-    ];
-
-    const rg = editor.addNewResponseComponent({ role: 'responseGroup' });
-    editor.addExistingResponseComponent({
-        role: 'text',
-        style: [{ key: 'className', value: 'mb-2' }],
-        content: generateLocStrings(
-            new Map([
-                ['nl-be', 'Optioneel in te vullen'],
-                ["fr-be", "Question facultative"],
-                ["de-be", "Optional einzutragen"],
-                ["en", "To be completed optionally"],
-            ])),
-    }, rg?.key);
-
-    editor.addExistingResponseComponent({
-        role: 'text',
-        style: [{ key: 'className', value: 'mb-1 fw-bold' }, { key: 'variant', value: 'h5' }],
-        content: generateLocStrings(
-            new Map([
-                ["nl-be", "Regelmatig handen wassen/desinfecteren"],
-                ["fr-be", "Se laver/se désinfecter régulièrement les mains"],
-                ["de-be", "Regelmäßig die Hände waschen/desinfiziren"],
-                ['en', 'Regularly wash or disinfect hands'],
-            ])),
-    }, rg?.key);
-    editor.addExistingResponseComponent(initLikertScaleItem(likertScaleKey + '_1', likertOptions), rg?.key);
-
-    editor.addExistingResponseComponent({
-        role: 'text',
-        style: [{ key: 'className', value: 'mb-1 border-top border-1 border-grey-7 pt-1 mt-2 fw-bold' }, { key: 'variant', value: 'h5' }],
-        content: generateLocStrings(
-            new Map([
-                ["nl-be", "Niezen en hoesten in uw elleboog"],
-                ["fr-be", "Éternuer et tousser dans votre coude"],
-                ["de-be", "In Ihren Ellenbogen niesen und husten"],
-                ['en', 'Cough or sneeze into your elbow'],
-            ])),
-    }, rg?.key);
-    editor.addExistingResponseComponent(initLikertScaleItem(likertScaleKey + '_2', likertOptions), rg?.key);
-
-    editor.addExistingResponseComponent({
-        role: 'text',
-        style: [{ key: 'className', value: 'mb-1 border-top border-1 border-grey-7 pt-1 mt-2 fw-bold' }, { key: 'variant', value: 'h5' }],
-        content: generateLocStrings(
-            new Map([
-                ["nl-be", "Wegwerptissues gebruiken"],
-                ["fr-be", "Utiliser des mouchoirs jetables"],
-                ["de-be", "Einwegtaschentücher verwenden"],
-                ['en', 'Use a disposable tissue'],
-            ])),
-    }, rg?.key);
-    editor.addExistingResponseComponent(initLikertScaleItem(likertScaleKey + '_3', likertOptions), rg?.key);
-
-    editor.addExistingResponseComponent({
-        role: 'text',
-        style: [{ key: 'className', value: 'mb-1 border-top border-1 border-grey-7 pt-1 mt-2 fw-bold' }, { key: 'variant', value: 'h5' }],
-        content: generateLocStrings(
-            new Map([
-                ["nl-be", "Een mondmasker dragen buitenshuis"],
-                ["fr-be", "Porter un masque buccal à l'extérieur"],
-                ["de-be", "Außerhalb des Hauses einen Mund-Nasen-Schutz tragen"],
-                ['en', 'Wear a face mask outdoors'],
-            ])),
-    }, rg?.key);
-    editor.addExistingResponseComponent(initLikertScaleItem(likertScaleKey + '_4', likertOptions), rg?.key);
-
-    editor.addExistingResponseComponent({
-        role: 'text',
-        style: [{ key: 'className', value: 'mb-1 border-top border-1 border-grey-7 pt-1 mt-2 fw-bold' }, { key: 'variant', value: 'h5' }],
-        content: generateLocStrings(
-            new Map([
-                ["nl-be", "Een mondmasker dragen binnenshuis"],
-                ["fr-be", "Porter un masque buccal à l'intérieur"],
-                ["de-be", "Im Hause einen Mund-Nasen-Schutz tragen"],
-                ['en', 'Wear a face mask indoors'],
-            ])),
-    }, rg?.key);
-    editor.addExistingResponseComponent(initLikertScaleItem(likertScaleKey + '_5', likertOptions), rg?.key);
-
-    editor.addExistingResponseComponent({
-        role: 'text',
-        style: [{ key: 'className', value: 'mb-1 border-top border-1 border-grey-7 pt-1 mt-2 fw-bold' }, { key: 'variant', value: 'h5' }],
-        content: generateLocStrings(
-            new Map([
-                ["nl-be", "Handen schudden vermijden"],
-                ["fr-be", "Éviter de serrer les mains"],
-                ["de-be", "Hände schütteln vermeiden"],
-                ['en', 'Avoid shaking hands'],
-            ])),
-    }, rg?.key);
-    editor.addExistingResponseComponent(initLikertScaleItem(likertScaleKey + '_6', likertOptions), rg?.key);
-
-    editor.addExistingResponseComponent({
-        role: 'text',
-        style: [{ key: 'className', value: 'mb-1 border-top border-1 border-grey-7 pt-1 mt-2 fw-bold' }, { key: 'variant', value: 'h5' }],
-        content: generateLocStrings(
-            new Map([
-                ["nl-be", "Niet meer knuffelen of kus op de wang geven"],
-                ["fr-be", "Éviter les embrassades ou les bisous sur la joue"],
-                ["de-be", "Nicht mehr knutschen oder einen Kuss auf die Wange geben"],
-                ['en', 'Stop greeting by hugging and/or kissing on both cheeks'],
-            ])),
-    }, rg?.key);
-    editor.addExistingResponseComponent(initLikertScaleItem(likertScaleKey + '_7', likertOptions), rg?.key);
-
-    editor.addExistingResponseComponent({
-        role: 'text',
-        style: [{ key: 'className', value: 'mb-1 border-top border-1 border-grey-7 pt-1 mt-2 fw-bold' }, { key: 'variant', value: 'h5' }],
-        content: generateLocStrings(
-            new Map([
-                ["nl-be", "Gebruik van openbaar vervoer beperken"],
-                ["fr-be", "Limiter l'utilisation des transports en commun"],
-                ["de-be", "Die Nutzung des ÖPNV begrenzen"],
-                ['en', 'Limit your use of public transport'],
-            ])),
-    }, rg?.key);
-    editor.addExistingResponseComponent(initLikertScaleItem(likertScaleKey + '_8', likertOptions), rg?.key);
-
-
-    editor.addExistingResponseComponent({
-        role: 'text',
-        style: [{ key: 'className', value: 'mb-1 border-top border-1 border-grey-7 pt-1 mt-2 fw-bold' }, { key: 'variant', value: 'h5' }],
-        content: generateLocStrings(
-            new Map([
-                ["nl-be", "Drukke plaatsen mijden (supermarkt, cinema, stadium,...)"],
-                ["fr-be", "Éviter les lieux bondés (un supermarché, un cinéma, un stade, etc.)"],
-                ["de-be", "Belebte Plätze meiden (Supermarkt, Kino, Stadion, usw.)"],
-                ['en', 'Avoid busy places (supermarket, cinema, stadium)'],
-            ])),
-    }, rg?.key);
-    editor.addExistingResponseComponent(initLikertScaleItem(likertScaleKey + '_9', likertOptions), rg?.key);
-
-
-    editor.addExistingResponseComponent({
-        role: 'text',
-        style: [{ key: 'className', value: 'mb-1 border-top border-1 border-grey-7 pt-1 mt-2 fw-bold' }, { key: 'variant', value: 'h5' }],
-        content: generateLocStrings(
-            new Map([
-                ["nl-be", "Thuis blijven"],
-                ["fr-be", "Rester à la maison"],
-                ["de-be", "Zu Hause bleiben"],
-                ['en', 'Stay at home'],
-            ])),
-    }, rg?.key);
-    editor.addExistingResponseComponent(initLikertScaleItem(likertScaleKey + '_10', likertOptions), rg?.key);
-
-    editor.addExistingResponseComponent({
-        role: 'text',
-        style: [{ key: 'className', value: 'mb-1 border-top border-1 border-grey-7 pt-1 mt-2 fw-bold' }, { key: 'variant', value: 'h5' }],
-        content: generateLocStrings(
-            new Map([
-                ["nl-be", "Telewerken"],
-                ["fr-be", "Effectuer du télétravail"],
-                ["de-be", "Telearbeit oder erhöhen Sie die Anzahl der Telearbeitstage"],
-                ['en', 'Telework or increase your number of telework days'],
-            ])),
-    }, rg?.key);
-    editor.addExistingResponseComponent(initLikertScaleItem(likertScaleKey + '_11', likertOptions), rg?.key);
-
-    editor.addExistingResponseComponent({
-        role: 'text',
-        style: [{ key: 'className', value: 'mb-1 border-top border-1 border-grey-7 pt-1 mt-2 fw-bold' }, { key: 'variant', value: 'h5' }],
-        content: generateLocStrings(
-            new Map([
-                ["nl-be", "Niet meer naar het buitenland reizen"],
-                ["fr-be", "Ne plus voyager à l'étranger"],
-                ["de-be", "Nicht mehr ins Ausland reisen!"],
-                ['en', 'Avoid travel outside your own country or region'],
-            ])),
-    }, rg?.key);
-    editor.addExistingResponseComponent(initLikertScaleItem(likertScaleKey + '_12', likertOptions), rg?.key);
-
-    editor.addExistingResponseComponent({
-        role: 'text',
-        style: [{ key: 'className', value: 'mb-1 border-top border-1 border-grey-7 pt-1 mt-2 fw-bold' }, { key: 'variant', value: 'h5' }],
-        content: generateLocStrings(
-            new Map([
-                ["nl-be", "Boodschappen aan huis laten leveren (door winkel/familie/vriend)"],
-                ["fr-be", "Se faire livrer ses achats à domicile (par un magasin/un membre de la famille/un ami)"],
-                ["de-be", "Aufträge nach Hause liefern lassen (durch das Geschäft/die Familie/Freunde)"],
-                ['en', 'Have your food/shopping delivered by a store or a friend/family member'],
-            ])),
-    }, rg?.key);
-    editor.addExistingResponseComponent(initLikertScaleItem(likertScaleKey + '_13', likertOptions), rg?.key);
-
-    editor.addExistingResponseComponent({
-        role: 'text',
-        style: [{ key: 'className', value: 'mb-1 border-top border-1 border-grey-7 pt-1 mt-2 fw-bold' }, { key: 'variant', value: 'h5' }],
-        content: generateLocStrings(
-            new Map([
-                ["nl-be", "Ontmoetingen met familie en vrienden beperken"],
-                ["fr-be", "Limiter les rencontres avec la famille et les amis"],
-                ["de-be", "Treffen mit Familie und Freunden einschränken und begrenzen"],
-                ['en', 'Avoid seeing friends and family'],
-            ])),
-    }, rg?.key);
-    editor.addExistingResponseComponent(initLikertScaleItem(likertScaleKey + '_14', likertOptions), rg?.key);
-
-    editor.addExistingResponseComponent({
-        role: 'text',
-        style: [{ key: 'className', value: 'mb-1 border-top border-1 border-grey-7 pt-1 mt-2 fw-bold' }, { key: 'variant', value: 'h5' }],
-        content: generateLocStrings(
-            new Map([
-                ["nl-be", "Ontmoetingen met mensen 65+ jaar of met chronische aandoeningen beperken"],
-                ["fr-be", "Limiter les rencontres avec les personnes âgées de plus de 65 ans ou souffrant de maladies chroniques"],
-                ["de-be", "Treffen mit Menschen im Alter von 65+ Jahren oder mit chronischen Erkrankungen begrenzen"],
-                ['en', 'Avoid being in contact with people over 65 years old or with a chronic disease'],
-            ])),
-    }, rg?.key);
-    editor.addExistingResponseComponent(initLikertScaleItem(likertScaleKey + '_15', likertOptions), rg?.key);
-
-    editor.addExistingResponseComponent({
-        role: 'text',
-        style: [{ key: 'className', value: 'mb-1 border-top border-1 border-grey-7 pt-1 mt-2 fw-bold' }, { key: 'variant', value: 'h5' }],
-        content: generateLocStrings(
-            new Map([
-                ["nl-be", "Contact met kinderen beperken"],
-                ["fr-be", "Limiter les contacts avec les enfants"],
-                ["de-be", "Kontakt mit Kindern begrenzen"],
-                ['en', 'Avoid being in contact with children'],
-            ])),
-    }, rg?.key);
-    editor.addExistingResponseComponent(initLikertScaleItem(likertScaleKey + '_16', likertOptions), rg?.key);
-
-    // VALIDATIONs
-    // None
 
     return editor.getItem();
 }
